Move enemies to their own module; use Enemy base class
