+++ characters.py
@@ -1,5 +1,61 @@
 
 
+#accuracy
+def accuracy(accuracy, attacker, target):
+    """
+    Determines whether an attack hits the opponent.
+    Can be changed by buffs or debuffs (statuses)
+    """
+    if target.has_status('Phantom'):
+        accuracy -= 10
+    if attacker.has_status('Nightfall'):
+        accuracy += 20
+    if accuracy <= 0:
+        return False
+    elif accuracy >= 100:
+        return True
+    hit = random.choice([True] * accuracy + [False] * (100 - accuracy))
+    if hit:
+        return True
+    else:
+        return False
+
+
+#Status
+    with open("statuses.json", "r") as f:
+        statuses = json.load(f)
+
+
+def infiltrated(damage):
+    """
+    Increases the damage taken by 10%
+    """
+    return abs(damage * 110 / 100)
+
+
+def instinct(damage):
+    """
+    Increases damage dealt by 30%
+    """
+    return abs(damage * 130 / 100)
+
+#Win and lose conditions
+def is_defeat(players: list) -> bool:
+    """
+    Returns True if user lost, else returns False
+    """
+    if len(players) == 0:
+        return True
+    return False
+
+def is_victory(enemies: list) -> bool:
+    """
+    Returns True if user won, else returns False
+    """
+    if len(enemies) == 0:
+        return True
+    return False
+
 
 class Character:
     """Base class for game characters.
